Move delete image action to imageSlice
