+++ src/features/loupe/FullSizeImage.jsx
@@ -7,7 +7,8 @@ import { styled } from '../../theme/stitches.config';
 import { selectUserUsername, selectUserCurrentRoles } from '../user/userSlice';
 import { hasRole, WRITE_OBJECTS_ROLES, WRITE_IMAGES_ROLES } from '../../auth/roles';
 import { drawBboxStart, selectIsDrawingBbox} from './loupeSlice';
-import { selectWorkingImages, labelsValidated, markedEmpty, deleteImages } from '../review/reviewSlice';
+import { selectWorkingImages, labelsValidated, markedEmpty } from '../review/reviewSlice';
+import { deleteImages } from '../images/imagesSlice';
 import { Image } from '../../components/Image';
 import BoundingBox from './BoundingBox';
 import DrawBboxOverlay from './DrawBboxOverlay';
